deconstructed fYFD
